add 10

+++ app/src/main/java/scut/carson_ho/shootatoffer/Exam_10.java
@@ -16,7 +16,7 @@ public class Exam_10 {
     /**
      * 递归实现
      */
-    public static long Fibonacci(int n){
+    public static int Fibonacci(int n){
         if(n<=1)
             return n;
         return Fibonacci(n-1)+Fibonacci(n-2);
@@ -25,16 +25,18 @@ public class Exam_10 {
     /**
      * 循环实现
      */
-    public static long Fibonacci2(int n){
+    public static int Fibonacci2(int n){
         if(n<=1)
             return n;
-        long fibOne = 1;
-        long fibTwo = 0;
-        long sum = 0;
+        
+        int fibOne = 1;
+        int fibTwo = 0;
+        int sum = 0;
+
         for (int i = 2; i <= n; i++) {
-            sum = fibOne+fibTwo;
+            sum = fibOne + fibTwo;
             fibTwo = fibOne;
-            fibOne = (int) sum;
+            fibOne = sum;
         }
         return sum;
     }
@@ -43,16 +45,21 @@ public class Exam_10 {
      * 青蛙跳台阶（基础）
      */
     public int jumpFloor(int number) {
-        if(number<=2)
+
+        if(number <=2 )
             return number;
+
         int jumpone=2; // 离所求的number的距离为1步的情况，有多少种跳法
         int jumptwo=1; // 离所求的number的距离为2步的情况，有多少种跳法
         int sum=0;
-        for(int i=3;i<=number;i++){
-            sum=jumptwo+jumpone;
-            jumptwo=jumpone;
-            jumpone=sum;
+
+        for( int i=3;i <= number; i++ ){
+
+            sum = jumptwo + jumpone;
+            jumptwo = jumpone;
+            jumpone= sum;
         }
+
         return sum;
     }
 
@@ -60,13 +67,18 @@ public class Exam_10 {
      * 青蛙跳台阶（变式）
      */
     public int jumpFloor2(int num) {
-        if(num<=2)
+        if( num <= 2)
             return num;
+
         int jumpone=2; // 前面一级台阶的总跳法数
+
         int sum=0;
-        for(int i=3;i<=num;i++){
+
+        for(int i=3; i<=num ;i++){
+
             sum = 2*jumpone;
             jumpone = sum;
+
         }
         return sum;
     }
